refactor: 로그 관련 코드 모듈로 분리
