+++ backend/src/app.py
@@ -2,6 +2,9 @@
 from fastapi import FastAPI, Request
 from routers import files, info, upload, chatbot, route_test  # 라우터 모듈 임포트
 import logging
+import web_utils.rag_init
+import web_utils.log_init
+from web_utils.log_init import setup_logging, log_middleware
 
 app = FastAPI(debug=True)
 
@@ -12,30 +15,12 @@ app.include_router(upload.router, prefix="/api/fileserver", tags=["Upload"])
 app.include_router(chatbot.router, prefix="/api/chatbot", tags=["chatbot"])
 app.include_router(route_test.router, prefix="/api/route_test", tags=["route_test"])
 
-# 전역 로그 제어
-# level=logging.INFO -> logger.info 출력
-# level=loggin.DEBUG -> logger.info, logger.debug 둘다 출력
-logger = logging.getLogger(__name__)
-logging.getLogger().setLevel(logging.INFO)
-logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s")
-http_debug = False
-@app.middleware("http")
-async def log_request(request: Request, call_next):
-    if http_debug:  # http_debug가 True일 때만 로그를 출력
-        # 요청 본문 출력
-        try:
-            body = await request.body()
-            logger.info(f"Request body: {body.decode('utf-8', errors='replace')}")  # 오류가 있는 바이트는 'replace'로 대체
-        except UnicodeDecodeError as e:
-            logger.error(f"Failed to decode request body: {e}")
-        
-        # 요청 헤더 출력
-        headers = dict(request.headers)
-        logger.info(f"Request headers: {headers}")
-    
-    # 요청을 처리하고, 응답을 반환
-    response = await call_next(request)
-    return response
+# 전역 로그 설정
+logger = setup_logging(level=logging.INFO)
+
+# HTTP 요청 로그 미들웨어 추가
+# http_debug=True 로 설정하는 경우, 디버그 메세지로 web request 를 확인 가능
+log_middleware(app, logger, http_debug=False)
 
 
 if __name__ == "__main__":
